+++ backend/canisters/community/api/src/queries/selected_updates_v2.rs
@@ -1,7 +1,7 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use ts_export::ts_export;
-use types::{CommunityMember, TimestampMillis, UserGroupDetails, UserId, VersionedRules};
+use types::{BotGroupDetails, CommunityMember, TimestampMillis, UserGroupDetails, UserId, VersionedRules};
 
 #[ts_export(community, selected_updates)]
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -26,6 +26,8 @@ pub struct SuccessResult {
     pub last_updated: TimestampMillis,
     pub members_added_or_updated: Vec<CommunityMember>,
     pub members_removed: Vec<UserId>,
+    pub bots_added_or_updated: Vec<BotGroupDetails>,
+    pub bots_removed: Vec<UserId>,
     pub blocked_users_added: Vec<UserId>,
     pub blocked_users_removed: Vec<UserId>,
     pub invited_users: Option<Vec<UserId>>,