+++ backend/libraries/group_community_common/src/lib.rs
@@ -1,4 +1,5 @@
 mod achievements;
+mod bots;
 mod expiring_member_actions;
 mod expiring_members;
 mod member;
@@ -7,6 +8,7 @@ mod pending_payments_queue;
 mod user_cache;
 
 pub use achievements::*;
+pub use bots::*;
 pub use expiring_member_actions::*;
 pub use expiring_members::*;
 pub use member::*;